add: Add Function to Toogle Like in Comments in Like Controller

+++ chitraverse_backend/src/controllers/like.controller.js
@@ -16,23 +16,72 @@ const toogleVideoLike = asyncHandler(async (req, res) => {
     throw new apiError(400, "Unauthorized User");
   }
 
-  const existingLike = await Like.findOne({ video: videoId, likedBy: userId });
-
-  if (existingLike) {
-    await Like.findByIdAndDelete(existingLike._id);
-    return res
+  try {
+    const existingLike = await Like.findOne({ video: videoId, likedBy: userId });
+  
+    if (existingLike) {
+      await Like.findByIdAndDelete(existingLike._id);
+      return res
+        .status(200)
+        .json(new apiResponse(200, "Video Unliked Successfully"));
+    } else {
+      const newLike = await Like.create({
+        video: videoId,
+        likedBy: userId,
+      });
+  
+      return res
       .status(200)
-      .json(new apiResponse(200, "Video Unliked Successfully"));
-  } else {
-    const newLike = await Like.create({
-      video: videoId,
-      likedBy: userId,
-    });
-
-    return res
-    .status(200)
-    .json(new apiResponse(200, newLike, "Video Liked Successfully"));
+      .json(new apiResponse(200, newLike, "Video Liked Successfully"));
+    }
+  } catch (error) {
+    console.log(400,"Error while toogling video like", error);
+    throw new apiError(400, "Something went wrong while toogling video like");
   }
 });
 
-export {toogleVideoLike}
+const toogleCommentLike = asyncHandler(async(req, res) =>{
+  const {id:commentId} = req.params;
+  
+  if(!commentId){
+    throw new apiError(400, "Invalid or Missing Comment Id");
+  }
+
+  const userId = req.user._id;
+
+  if(!userId){
+    throw new apiError(400, "Authorization Error");
+  }
+
+  try {
+     const existingLike = await Like.findOne({comment:commentId, likedBy:userId});
+
+     if(existingLike){
+      await Like.findByIdAndDelete(existingLike._id);
+      return res
+      .status(200)
+      .json(new apiResponse(200,"Comment Unliked Successfully"));
+     }
+     else{
+      const newLike = await Like.create({
+        comment:commentId,
+        likedBy:userId
+      });
+
+      if(!newLike){
+        throw new apiError(400, "Error while liking comment");
+      }
+
+      return res
+      .status(200)
+      .json(new apiError(200, newLike, "Comment Liked Successfully"));
+     }
+  } catch (error) {
+    console.log("Error while toogling comment like", error);
+    throw new apiError(400, "Error while toogling comment like" );
+  }
+
+  
+})
+
+export {toogleVideoLike, toogleCommentLike}
